Add a single entry point for the app
