+++ mobile/__tests__/index.android.js
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import Index from '../index.android.js';
+import Index from '../src';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
