+++ mobile/__tests__/index.ios.js
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import Index from '../index.ios.js';
+import Index from '../src';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
